전남대 BE_이상현 2단계 - 관리자 화면 구현 (#129)
* refactor: change DTO class to record class

* feats: implement all products GET API

* docs: add admin page docs to README

* refactor: change price type to long

* feats: 관리자용 페이지, 상품 조회 기능 구현

* feats: 상품 추가 기능 구현

* feats: 상품 수정 기능 구현

* feats: 상품 삭제 기능 구현

+++ src/main/java/gift/entity/Product.java
@@ -3,10 +3,10 @@ package gift.entity;
 public class Product {
     private Long id;
     private String name;
-    private int price;
+    private long price;
     private String imageUrl;
 
-    public Product(String name, int price, String imageUrl) {
+    public Product(String name, long price, String imageUrl) {
         this.name = name;
         this.price = price;
         this.imageUrl = imageUrl;
@@ -24,7 +24,7 @@ public class Product {
         return name;
     }
 
-    public int getPrice() {
+    public long getPrice() {
         return price;
     }
 
@@ -32,7 +32,7 @@ public class Product {
         return imageUrl;
     }
 
-    public void updateProduct(String name, int price, String imageUrl) {
+    public void updateProduct(String name, long price, String imageUrl) {
         this.name = name;
         this.price = price;
         this.imageUrl = imageUrl;
